requirements.txt
ccxt>=4.0.0
pyyaml>=6.0
pandas>=2.0.0
rich>=13.0.0
questionary>=2.0.0
python-dotenv>=1.0.0
